- Use UBO for unpack hdr - Fixed god ray's center position - DAELoader now loads related texture files - Fixed skinning code

+++ projects/PathosEngine/src/pathos/console.cpp
@@ -84,7 +84,7 @@ namespace pathos {
 	Label* ConsoleWindow::addLine(const char* text) {
 		int size = MultiByteToWideChar(CP_ACP, 0, text, -1, NULL, NULL);
 		std::vector<wchar_t> buffer(size, 0);
-		MultiByteToWideChar(CP_ACP, 0, text, strlen(text)+1, buffer.data(), size);
+		MultiByteToWideChar(CP_ACP, 0, text, static_cast<int>(strlen(text) + 1), buffer.data(), size);
 		return addLine(buffer.data());
 	}
